+++ app/src/main/java/com/example/moviesearch/App.kt
@@ -4,58 +4,28 @@ import android.app.Application
 import com.example.moviesearch.data.ApiConstants
 import com.example.moviesearch.data.MainRepository
 import com.example.moviesearch.data.TmdbApi
+import com.example.moviesearch.di.DI
 import com.example.moviesearch.domain.Interactor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
+import org.koin.android.ext.koin.androidContext
+import org.koin.android.ext.koin.androidLogger
+import org.koin.core.context.startKoin
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import java.util.concurrent.TimeUnit
 
 class App : Application() {
-    lateinit var repo: MainRepository
-    lateinit var interactor: Interactor
-    lateinit var retrofitService: TmdbApi
 
     override fun onCreate() {
         super.onCreate()
-        //Инициализируем экземпляр App, через который будем получать доступ к остальным переменным
-        instance = this
-        //Инициализируем репозиторий
-        repo = MainRepository()
-
-        //Создаём кастомный клиент
-        val okHttpClient = OkHttpClient.Builder()
-            //Настраиваем таймауты для медленного интернета
-            .callTimeout(30, TimeUnit.SECONDS)
-            .readTimeout(30, TimeUnit.SECONDS)
-            //Добавляем логгер
-            .addInterceptor(HttpLoggingInterceptor().apply {
-                if (BuildConfig.DEBUG) {
-                    level = HttpLoggingInterceptor.Level.BASIC
-                }
-            })
-            .build()
-//Создаем Ретрофит
-        val retrofit = Retrofit.Builder()
-            //Указываем базовый URL из констант
-            .baseUrl(ApiConstants.BASE_URL)
-            //Добавляем конвертер
-            .addConverterFactory(GsonConverterFactory.create())
-            //Добавляем кастомный клиент
-            .client(okHttpClient)
-            .build()
-//Создаем сам сервис с методами для запросов
-        retrofitService = retrofit.create(TmdbApi::class.java)
-//Инициализируем интерактор
-        interactor = Interactor(repo, retrofitService)
-        //Инициализируем интерактор
-        interactor = Interactor(repo, retrofitService)
-    }
-
-    companion object {
-        //Здесь статически хранится ссылка на экземпляр App
-        lateinit var instance: App
-            //Приватный сеттер, чтобы нельзя было в эту переменную присвоить что-либо другое
-            private set
+        startKoin {
+            //Прикрепляем контекст
+            androidContext(this@App)
+            //(Опционально) подключаем зависимость
+            androidLogger()
+            //Инициализируем модули
+            modules(listOf(DI.mainModule))
+        }
     }
 }
