+++ app/src/main/java/com/example/moviesearch/viewmodel/HomeFragmentViewModel.kt
@@ -5,11 +5,13 @@ import androidx.lifecycle.ViewModel
 import com.example.moviesearch.App
 import com.example.moviesearch.domain.Film
 import com.example.moviesearch.domain.Interactor
+import org.koin.core.KoinComponent
+import org.koin.core.inject
 
-class HomeFragmentViewModel : ViewModel() {
-    val filmsListLiveData = MutableLiveData<List<Film>>()
-    //Инициализируем интерактор
-    private var interactor: Interactor = App.instance.interactor
+class HomeFragmentViewModel : ViewModel(), KoinComponent {
+        val filmsListLiveData:  MutableLiveData<List<Film>> = MutableLiveData()
+        //Инициализируем интерактор
+        private val interactor: Interactor by inject()
     init {
         interactor.getFilmsFromApi(1, object : ApiCallback {
             override fun onSuccess(films: List<Film>) {
